Get rid of isKerberos, introducing canRetry() and relogin() methods
The logic is still kind of weird, but I think an improvement.
