+++ hbase-client/src/main/java/org/apache/hadoop/hbase/security/provider/AbstractSaslClientAuthenticationProvider.java
@@ -19,7 +19,6 @@ package org.apache.hadoop.hbase.security.provider;
 
 import org.apache.hadoop.hbase.HBaseInterfaceAudience;
 import org.apache.hadoop.io.Text;
-import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.yetus.audience.InterfaceAudience;
 import org.apache.yetus.audience.InterfaceStability;
 
@@ -57,16 +56,5 @@ public abstract class AbstractSaslClientAuthenticationProvider implements
       return this.hashCode() == o.hashCode();
     }
     return false;
-  }
-
-  @Override
-  public UserGroupInformation unwrapUgi(UserGroupInformation ugi) {
-    // Unwrap the UGI with the real user when we're using Kerberos auth
-    if (isKerberos() && ugi != null && ugi.getRealUser() != null) {
-      return ugi.getRealUser();
-    }
-
-    // Otherwise, use the UGI we were given
-    return ugi;
   }
 }
