+++ hbase-server/src/test/java/org/apache/hadoop/hbase/security/provider/TestCustomSaslAuthenticationProvider.java
@@ -265,11 +265,6 @@ public class TestCustomSaslAuthenticationProvider {
     public UserInformation getUserInfo(UserGroupInformation user) {
       return null;
     }
-
-    @Override
-    public boolean isKerberos() {
-      return false;
-    }
   }
 
   /**
